worker_services/requirements.txt
pika==1.3.2
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
